+++ modules/apps/object/object-scripting-impl/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Object Runtime Scripting Implementation
-Bundle-SymbolicName: com.liferay.object.runtime.scripting.impl
-Bundle-Version: 1.0.2
+Bundle-Name: Liferay Object Scripting Implementation
+Bundle-SymbolicName: com.liferay.object.scripting.impl
+Bundle-Version: 1.0.0
 DynamicImport-Package: *
 -dsannotations-options: inherit
